requirements.txt
future==1.0.0
iso8601==2.1.0
paho-mqtt==2.1.0
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.2
serial==0.0.97
